s3/django_storage/backend/requirements.txt
asgiref==3.9.2
boto3==1.40.60
botocore==1.40.60
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
jmespath==1.0.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
